cleanup + tests
